sql/parser: move the ColumnType constructors and pre-allocs to their own file
